// src/types/api/walletType.ts
/**
 * Kriptonio Private API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: private
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type WalletType = 'Eoa' | 'SmartWallet';

export const WalletType = {
    Eoa: 'Eoa' as WalletType,
    SmartWallet: 'SmartWallet' as WalletType
};